+++ src/modules/scheduler/decorator/index.ts
@@ -1,0 +1,2 @@
+export * from './job.decorator';
+export * from './scheduler.decorator';