+++ src/modules/scheduler/decorator/scheduler.decorator.ts
@@ -1,0 +1,7 @@
+import { BindingScope, injectable } from '@loopback/context';
+
+export const Scheduler = () => {
+  return injectable(binding => {
+    binding.inScope(BindingScope.SINGLETON).tag('scheduler');
+  });
+};